(Sub) Updated Airing Schedule
(Sub) Added Missing Episode 1 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 2 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 3 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 4 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 5 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 6 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 7 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Sub) Added Missing Episode 8 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V