(Dub) Added New Episode(s)
(Dub) Added Episode 1 for Plus-Sized Elf
(Dub) Added Episode 4 for How I Attended an All-Guy's Mixer
(Dub) Added Episode 5 for Loner Life in Another World
(Dub) Added (multi-header) release Episode 12 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V
(Dub) Added (multi-header) release Episode 13 for Is It Wrong To Try To Pick Up Girls in a Dungeon? V