fix: (Dub) batch episode predictions
We shouldn't predict any batch drops as these are likely infrequent instead of weekly, like One Piece and Pokemon.
A buffer of 4 episodes is used since sometimes we can have weekly of 2 episode batches or an initial batch of 1-4 as an example.
If the batch drop is more than 4 and its not a single batch drop, its likely the entire series so we still don't need to predict...