(Dub) (Added/Removed) Episode(s)
(Dub) Removed Episode 5 of GINTAMA - Mr. Ginpachi's Zany Class due to a correction in the airing date
(Dub) Removed Episode 2 of Isekai Quartet 3 due to a correction in the airing date
(Dub) Removed Episode 4 of A Gatherer's Adventure in Isekai due to a correction in the airing date
(Dub) Added Episode 5 for GINTAMA - Mr. Ginpachi's Zany Class
(Dub) Added Episode 2 for Isekai Quartet 3
(Dub) Added Episode 4 for A Gatherer's Adventure in Isekai