(Sub) Updated Airing Schedule
(Sub) Updated Episode 9 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 8 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 7 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 6 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 5 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 4 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 3 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 2 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.
(Sub) Updated Episode 1 for Chitose Is in the Ramune Bottle to correct its idMal, format, and duration.